% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICONx.R
\name{bayes_base}
\alias{bayes_base}
\title{Creat BayesPrism base}
\usage{
bayes_base(
  bk.dat,
  sc.dat,
  cell.type.labels,
  species = "hs",
  input.type = "count.matrix",
  gene.group = c("Rb", "Mrp", "other_Rb", "chrM", "MALAT1", "chrX", "chrY"),
  gene.type = "protein_coding"
)
}
\arguments{
\item{bk.dat}{is a matrix with rows for gene names and columns for sample names}

\item{sc.dat}{is a matrix with rows for gene names and columns for cell names}

\item{cell.type.labels}{is an n*2 matrix, the first column is the cell name of sc.dat, and the second column is the cell subtype}
}
\value{
BayesPrism base
}
\description{
Creat BayesPrism base
}
\examples{
bayes_base(bulk.samples,sc.dat,subtype)
}
