% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICONx.R
\name{BayesPrism_base}
\alias{BayesPrism_base}
\title{Create BayesPrism base}
\usage{
BayesPrism_base(
  bk.dat,
  sc.dat,
  cell.type.labels,
  species = "hs",
  input.type = "count.matrix",
  gene.group = c("Rb", "Mrp", "other_Rb", "chrM", "MALAT1", "chrX", "chrY"),
  gene.type = "protein_coding"
)
}
\arguments{
\item{bk.dat}{is a matrix with rows for gene names and columns for sample names}

\item{sc.dat}{is a matrix with rows for gene names and columns for cell names}

\item{cell.type.labels}{is an n*1 matrix, the rowname is the cell name of sc.dat, and the first column is the cell subtype}

\item{species}{character string indicating species, defaults to "hs" (human)}

\item{input.type}{character string specifying input type, defaults to "count.matrix"}

\item{gene.group}{character vector of gene groups to filter}

\item{gene.type}{character string specifying gene type, defaults to "protein_coding"}
}
\value{
BayesPrism base
}
\description{
Create BayesPrism base
}
\examples{
BayesPrism_base(bulk.samples, sc.dat, subtype)
}
