% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICONx.R
\name{generate_signature_matrix}
\alias{generate_signature_matrix}
\title{generate_signature_matrix}
\usage{
generate_signature_matrix(cell_types)
}
\arguments{
\item{cell_types}{can enter in nk, b, cd4, cd8, fibroblast, five kinds of macrophage cell types}
}
\description{
generate_signature_matrix
}
