% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICONx.R
\name{monocle3_base}
\alias{monocle3_base}
\title{Creat monocle3 base}
\usage{
monocle3_base(expression_matrix, cell_metadata)
}
\arguments{
\item{expression_matrix}{is a matrix with rows for gene names and columns for cell names}

\item{cell_metadata}{is an n*2 matrix, the first column is the cell name of sc.dat, and the second column is the cell subtype}
}
\value{
monocle3 base
}
\description{
Creat monocle3 base
}
\examples{
monocle3_base <- monocle3_base(sc.dat,subtype)
}
