% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICONx.R
\name{MuSiC2_base}
\alias{MuSiC2_base}
\title{Creat MuSiC2 base}
\usage{
MuSiC2_base(
  sc.sce,
  bulk,
  cell.type.labels,
  prop_r = 0.1,
  eps_c = 0.05,
  expr_low = 20,
  cutoff_expr = 0.05,
  maxiter = 200,
  markers = NULL,
  cell_size = NULL,
  centered = FALSE,
  ct_cov = FALSE,
  normalize = FALSE,
  cutoff_fc = 2,
  eps_r = 0.01,
  n_resample = 20,
  sample_prop = 0.5,
  cutoff_c = 0.05,
  cutoff_r = 0.01,
  iter_max = 1000,
  nu = 1e-04,
  eps = 0.01,
  non.zero = FALSE
)
}
\arguments{
\item{sc.sce}{is a matrix with rows for gene names and columns for cell names}

\item{bulk}{is a matrix with rows for gene names and columns for sample names}

\item{cell.type.labels}{is an n*2 matrix, the first column is the cell name of sc.dat, and the second column is the cell subtype}
}
\value{
MuSiC2 base
}
\description{
Creat MuSiC2 base
}
