% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECEPTICONx.R
\name{run_DECEPTICONx}
\alias{run_DECEPTICONx}
\title{run_DECEPTICONx}
\usage{
run_DECEPTICONx(
  bulk.samples,
  sc.dat,
  subtype,
  cell_types = c("nk", "b", "cd4", "cd8", "fibroblast", "macrophage"),
  RUNpath,
  light = TRUE,
  custom.template = FALSE,
  custom.signature = TRUE,
  single.cell = FALSE,
  single.cell.data = NULL
)
}
\arguments{
\item{bulk.samples}{is a matrix with rows for gene names and columns for sample names}

\item{sc.dat}{is a matrix with rows for gene names and columns for cell names}

\item{subtype}{is an n*2 matrix, the first column is the cell name of sc.dat, and the second column is the cell subtype}

\item{cell_types}{can enter in nk, b, cd4, cd8, fibroblast, five kinds of macrophage cell types}

\item{light}{set TRUE for light version of DECEPTICON, running time reduced.}

\item{custom.signature}{set TRUE for custom signature matrix.}

\item{signature.matrix}{When custom.signature is set to TRUE, need to enter expression template(s).}
}
\description{
run_DECEPTICONx
}
